卒論が区切りついたのでコミット

+++ src/exp_env/reservoir/bit_esn.py
@@ -82,7 +82,7 @@ class BitESNModule(nn.Module,BaseReservoirModule):
         if self.initial_state is not None:
             self.log_state = [self.initial_state]
         else:
-            self.log_state = [torch.zeros(self.out_features).to(self.w_res.device)]
+            self.log_state = [torch.zeros(self.out_features).to(self.w_res.device).to(self.w_res.dtype)]
     
     def forward(self,X:torch.Tensor,one_step_mode=False)->torch.Tensor:
         if one_step_mode:
